moved show decision methods into api folder since they're not used multiple times
